tests/krate/publish: Move more tests to the `max_size` module
